Remove subprojects
mimic-core needs libpcre2-8, but it does not have
meson support (and I am not going to write it now).

Also we would need to set PKG_CONFIG_PATH in a script
anyway before building to a custom prefix, because
meson does not look into the prefix when searching
for dependencies. So given that we need a script anyway
let's let the script do the dependency installation
as well.